pyproject.toml
[project]
name = "vinyl-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.128.0",
    "google-auth>=2.45.0",
    "gspread>=6.2.1",
    "jinja2>=3.1.6",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyprojroot>=0.3.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "python3-discogs-client>=2.8",
    "uvicorn>=0.40.0",
]

[tool.setuptools]
packages = ["vinyl_recorder"]
